requirements.txt
numpy
soundfile
librosa
torch
transformers
tqdm
sherpa-onnx
pyyaml